\docType{data}
\name{hidden_aliases}

\alias{+,_p_itk__simple__Image,_p_itk__simple__Image-method}
\alias{+,numeric,_p_itk__simple__Image-method}
\alias{+,_p_itk__simple__Image,numeric-method}

\alias{-,_p_itk__simple__Image,_p_itk__simple__Image-method}
\alias{-,numeric,_p_itk__simple__Image-method}
\alias{-,_p_itk__simple__Image,numeric-method}

\alias{*,_p_itk__simple__Image,_p_itk__simple__Image-method}
\alias{*,numeric,_p_itk__simple__Image-method}
\alias{*,_p_itk__simple__Image,numeric-method}

\alias{^,_p_itk__simple__Image,_p_itk__simple__Image-method}
\alias{^,numeric,_p_itk__simple__Image-method}
\alias{^,_p_itk__simple__Image,numeric-method}

\alias{\%\%,_p_itk__simple__Image,_p_itk__simple__Image-method}
\alias{\%\%,numeric,_p_itk__simple__Image-method}
\alias{\%\%,_p_itk__simple__Image,numeric-method}

\alias{\%/\%,_p_itk__simple__Image,_p_itk__simple__Image-method}
\alias{\%/\%,numeric,_p_itk__simple__Image-method}
\alias{\%/\%,_p_itk__simple__Image,numeric-method}

\alias{/,_p_itk__simple__Image,_p_itk__simple__Image-method}
\alias{/,numeric,_p_itk__simple__Image-method}
\alias{/,_p_itk__simple__Image,numeric-method}

\alias{abs,_p_itk__simple__Image-method}
\alias{acos,_p_itk__simple__Image-method}
\alias{cos,_p_itk__simple__Image-method}
\alias{asin,_p_itk__simple__Image-method}
\alias{sin,_p_itk__simple__Image-method}
\alias{atan,_p_itk__simple__Image-method}
\alias{tan,_p_itk__simple__Image-method}

\alias{sqrt,_p_itk__simple__Image-method}

\alias{exp,_p_itk__simple__Image-method}
\alias{log,_p_itk__simple__Image-method}
\alias{log10,_p_itk__simple__Image-method}
\alias{print,_p_itk__simple__Image-method}
\alias{show,_p_itk__simple__Image-method}


\alias{!,_p_itk__simple__Image-method}
\alias{|,_p_itk__simple__Image,_p_itk__simple__Image-method}
\alias{&,_p_itk__simple__Image,_p_itk__simple__Image-method}

\alias{[,ExternalReference-method}
\alias{[<-,ExternalReference-method}
\alias{print,ExternalReference-method}
\alias{show,ExternalReference-method}

\alias{print,C++Reference-method}
\alias{show,C++Reference-method}

\alias{length,SWIGArray-method}

\alias{!=,_p_itk__simple__Image,_p_itk__simple__Image-method}
\alias{!=,numeric,_p_itk__simple__Image-method}
\alias{!=,_p_itk__simple__Image,numeric-method}

\alias{==,_p_itk__simple__Image,_p_itk__simple__Image-method}
\alias{==,numeric,_p_itk__simple__Image-method}
\alias{==,_p_itk__simple__Image,numeric-method}

\alias{<=,_p_itk__simple__Image,_p_itk__simple__Image-method}
\alias{<=,numeric,_p_itk__simple__Image-method}
\alias{<=,_p_itk__simple__Image,numeric-method}

\alias{>=,_p_itk__simple__Image,_p_itk__simple__Image-method}
\alias{>=,numeric,_p_itk__simple__Image-method}
\alias{>=,_p_itk__simple__Image,numeric-method}

\alias{<,_p_itk__simple__Image,_p_itk__simple__Image-method}
\alias{<,numeric,_p_itk__simple__Image-method}
\alias{<,_p_itk__simple__Image,numeric-method}

\alias{>,_p_itk__simple__Image,_p_itk__simple__Image-method}
\alias{>,numeric,_p_itk__simple__Image-method}
\alias{>,_p_itk__simple__Image,numeric-method}

\alias{as.array,_p_itk__simple__Image-method}

\alias{hidden_aliases}

\title{Internal page for hidden aliases}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
